hw-05 avatars - patching successfully, needs to squish the avatar
