+++ models/user.js
@@ -24,6 +24,7 @@ const userScheme = new Schema(
       default: "starter",
     },
     token: { type: String, default: "" },
+    avatarURL: { type: String, required: true },
   },
   { versionKey: false }
 );
